작성자 - 윤형식 (도서)
작성일자 - 2020.07.03

완성한 내용
- 도서 추가기능, Dto에서 도서코드 삭제

진행중인 내용
- 도서 수정, 도서 삭제

+++ lightbrary/src/main/java/com/lightbrary/book/model/BookImageDto.java
@@ -6,7 +6,7 @@ public class BookImageDto {
 	private int no;
 	private String originalFileName;
 	private String storedFileName;
-	private int fileSize;
+	private long fileSize;
 	private Date createdDate;
 	
 	public BookImageDto() {
@@ -39,10 +39,10 @@ public class BookImageDto {
 	public void setStoredFileName(String storedFileName) {
 		this.storedFileName = storedFileName;
 	}
-	public int getFileSize() {
+	public long getFileSize() {
 		return fileSize;
 	}
-	public void setFileSize(int fileSize) {
+	public void setFileSize(long fileSize) {
 		this.fileSize = fileSize;
 	}
 	public Date getCreatedDate() {
